requirements.txt
absl-py==2.1.0
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
cffi==1.17.1
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
Flask==3.0.3
Flask-SocketIO==5.4.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
h11==0.14.0
itsdangerous==2.2.0
jax==0.4.34
jaxlib==0.4.34
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mediapipe==0.10.15
ml_dtypes==0.5.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.5
pycparser==2.22
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.4
scikit-learn==1.5.2
scipy==1.14.1
simple-websocket==1.0.0
six==1.16.0
sounddevice==0.5.0
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
triton==3.0.0
typing_extensions==4.12.2
Werkzeug==3.0.4
wsproto==1.2.0
